fix: :truck: Moved docs/lib.ts to lib/docs.ts
Issue: #129
